+++ lua/github-theme/group/editor.lua
@@ -4,6 +4,7 @@ function M.get(spec, config)
   local dark_sb = config.darken.sidebars.enable
   local hide_eof = config.hide_end_of_buffer
   local inactive = config.dim_inactive
+  local inv = config.inverse
   local trans = config.transparent
 
   local sb_bg
@@ -15,13 +16,7 @@ function M.get(spec, config)
     sb_bg = spec.bg1
   end
 
-  local P = spec.palette
-
-  -- TODO:
-  -- (1) Config: Add options.dim_inactive
-  -- (2) Config: Add options.inverse
-
-  -- local inv = config.inverse
+  local c = spec.palette
 
  -- stylua: ignore start
  local groups = {
@@ -41,7 +36,7 @@ function M.get(spec, config)
     -- TermCursor      = {}, -- cursor in a focused terminal
     -- TermCursorNC    = {}, -- cursor in an unfocused terminal
     ErrorMsg        = { fg = spec.diag.error }, -- error messages on the command line
-    VertSplit       = { fg = P.border.default }, -- the column separating vertically split windows
+    VertSplit       = { fg = c.border.default }, -- the column separating vertically split windows
     Folded          = { fg = spec.fg3, bg = spec.bg2 }, -- line used for closed folds
     FoldColumn      = { fg = spec.fg3 }, -- 'foldcolumn'
     SignColumn      = { fg = spec.fg3 }, -- column where |signs| are displayed
@@ -50,11 +45,10 @@ function M.get(spec, config)
     LineNr          = { fg = spec.fg0 }, -- Line number for ':number' and ':#' commands, and when 'number' or 'relativenumber' option is set.
     CursorLineNr    = { fg = spec.fg1 }, -- Like LineNr when 'cursorline' or 'relativenumber' is set for the cursor line.
 
-    MatchParen      = { fg = spec.fg1, bg = P.accent.muted }, -- The character under the cursor or just before it, if it is a paired bracket, and its match. |pi_paren.txt|
-    -- MatchParen      = { fg = spec.diag.warn, style = inv.match_paren and 'reverse,bold' or 'bold' }, -- The character under the cursor or just before it, if it is a paired bracket, and its match. |pi_paren.txt|
+    MatchParen      = { fg = spec.fg1, bg = c.accent.muted, style = inv.match_paren and 'reverse,bold' or 'bold' }, -- The character under the cursor or just before it, if it is a paired bracket, and its match. |pi_paren.txt|
 
     ModeMsg         = { fg = spec.diag.warn, style = 'bold' }, -- 'showmode' message (e.g., '-- INSERT -- ')
-    -- NOTE: This was commented out as there is an issue with seting this highlight group see issue #98
+    -- NOTE: This was commented out as there is an issue with setting this highlight group see issue #98
     -- MsgArea         = { fg = spec.fg2 }, -- Area for messages and cmdline
     -- MsgSeparator    = {}, -- Separator for scrolled messages, `msgsep` flag of 'display'
     MoreMsg         = { fg = spec.diag.info, style = 'bold' }, -- |more-prompt|
@@ -65,7 +59,7 @@ function M.get(spec, config)
     NormalNC        = { fg = spec.fg1, bg = (inactive and spec.bg0) or (trans and 'NONE') or spec.bg1 }, -- normal text in non-current windows
 
     NormalFloat     = { fg = spec.fg1, bg = spec.bg0 }, -- Normal text in floating windows.
-    FloatBorder     = { fg = P.border.default }, -- TODO
+    FloatBorder     = { fg = c.border.default }, -- TODO
     Pmenu           = { fg = spec.fg1, bg = spec.bg0 }, -- Popup menu: normal item.
     PmenuSel        = { bg = spec.sel1 }, -- Popup menu: selected item.
     PmenuSbar       = { link = 'Pmenu' }, -- Popup menu: scrollbar.
@@ -73,10 +67,8 @@ function M.get(spec, config)
     Question        = { link = 'MoreMsg' }, -- |hit-enter| prompt and yes/no questions
     QuickFixLine    = { link = 'CursorLine' }, -- Current |quickfix| item in the quickfix window. Combined with |hl-CursorLine| when the cursor is there.
 
-    Search          = { bg = spec.sel2 }, -- Last search pattern highlighting (see 'hlsearch').  Also used for similar items that need to stand out.
-    IncSearch       = { fg = P.black.bright, bg = P.orange }, -- 'incsearch' highlighting; also used for the text replaced with ':s///c'
-    -- Search          = inv.search and { style = 'reverse' } or { bg = spec.sel2 }, -- Last search pattern highlighting (see 'hlsearch').  Also used for similar items that need to stand out.
-    -- IncSearch       = inv.search and { style = 'reverse' } or { fg = c.black.bright, bg = c.orange }, -- 'incsearch' highlighting; also used for the text replaced with ':s///c'
+    Search          = inv.search and { style = 'reverse' } or { bg = spec.sel2 }, -- Last search pattern highlighting (see 'hlsearch').  Also used for similar items that need to stand out.
+    IncSearch       = inv.search and { style = 'reverse' } or { fg = c.black.bright, bg = c.orange }, -- 'incsearch' highlighting; also used for the text replaced with ':s///c'
 
     CurSearch       = { link = 'IncSearch'}, -- Search result under cursor (available since neovim >0.7.0 (https://github.com/neovim/neovim/commit/b16afe4d556af7c3e86b311cfffd1c68a5eed71f)).
     SpecialKey      = { link = 'NonText' }, -- Unprintable characters: text displayed differently from what it really is.  But not 'listchars' whitespace. |hl-Whitespace|
@@ -92,10 +84,8 @@ function M.get(spec, config)
     TabLineSel      = { fg = spec.bg1, bg = spec.fg3 }, -- tab pages line, active tab page label
     Title           = { fg = spec.syntax.builtin2, style = 'bold' }, -- titles for output from ':set all', ':autocmd' etc.
 
-    Visual          = { bg = spec.sel0 }, -- Visual mode selection
-    VisualNOS       = { link = 'visual' }, -- Visual mode selection when vim is 'Not Owning the Selection'.
-    -- Visual          = inv.visual and { style = 'reverse' } or { bg = spec.sel0 }, -- Visual mode selection
-    -- VisualNOS       = inv.visual and { style = 'reverse' } or { link = 'visual' }, -- Visual mode selection when vim is 'Not Owning the Selection'.
+    Visual          = inv.visual and { style = 'reverse' } or { bg = spec.sel0 }, -- Visual mode selection
+    VisualNOS       = inv.visual and { style = 'reverse' } or { link = 'visual' }, -- Visual mode selection when vim is 'Not Owning the Selection'.
 
     WarningMsg      = { fg = spec.diag.warn }, -- warning messages
     Whitespace      = { fg = spec.bg3 }, -- 'nbsp', 'space', 'tab' and 'trail' in 'listchars'